applications/portfoliomanager/pyproject.toml
[project]
name = "portfoliomanager"
version = "0.1.0"
description = "Portfolio prediction and construction service"
requires-python = "==3.12.10"
dependencies = [
    "internal",
    "fastapi>=0.116.1",
    "uvicorn>=0.35.0",
    "httpx>=0.27.0",
    "pandera[polars,pandas]>=0.26.0",
    "alpaca-py>=0.42.1",
]

[tool.uv]
package = true
src = ["src"]

[tool.uv.sources]
internal = { workspace = true }
